refactor: Change location of utils imports.
